Extract all xml expression logic to a partial file
